Continued build improvements
Continued improving clang-based building of Sailfish and
Salmon.  There seems to be an issue with Boost-1.55 and
Clang 3.4 --- it's fixed upstream, but necessitates a
new preprocessor define when building with clang.  I put
this into the build under clang temporarily, but it may break
things if the processor doesn't have 128-bit integer support.