Check SAM/BAM input for proper alignment pairs
Alignment-based salmon expects that the alignment records for
two ends of a paired-end read will be adjacent in the SAM/BAM
file.  Check for this condition.  Throw an error an exit if
we encounter an unpaired record, and report an annoying warning
and skip the record if the names of an adjacent pair of
alignment records do not match.  Think more about if this is
the best / most robust way to handle paired-end alignment
files.