Lots of work to all gcc/clang osx/linux
Updated to latest TBB --- had to get rid of a few rvalue pushes onto
a tbb_concurrent_queue  (not sure why; this should be supported)

On a very old assembler (like the on on Travis CI), we have to pass
-DNO_RTM to the build system b/c the newer compiler emits code that
the assember cannot generate instructions for (in this case for
restricted transactional memory)

Deal with some other edge cases in gcc vs. clang.