Significant commit, with some fixes / features
Fixed a bug in the writing of sampled alignments that could
cause a small number of alignments (the # of reads modulo the
mini-batch size --- 1000 by default) to be skipped.

Added a completely new error model to alignment-based salmon.
This model combines some ideas from the TIGAR model (i.e.
it's essentially a 1st order Markov Chain that allows us to
evaluate the probability of a given alignment --- the current
alignment state, taking into account the nucleotides and CIGAR
operations, are dependent upon the previous state) and the
eXpress model (i.e. unlike the model of TIGAR, there is spatial
heterogineity in the model so that e.g. the indel probability
varies depeneding on the relative position within the read).
This model has the benefit that it places no a priori bound on
the read length, but also takes into acount position specific
context information when evaluating the probability of an alignment.

A number of other smaller optimizations and fixes were also made.

This is the first commit in preparation for the v0.2.8 release.