Added PCA implementation (custom) using eigen
This commit rids Sailfish and Salmon (hopefully, once and for all)
of the dependency on the insanely finicky Shark library.  The
(very simple) custom PCA implementation (in the file PCA.hpp,
that was added in this commit) seems to achieve the same thing
as does Shark, whereas some other PCA implementations I tried before
didn't.  Shark; be gone!