Buf fix
In the (rare) event that one is quantifying with
paired-end reads and there are fewer than 50,000
uniquely mapping reads (so that the fragment length
prior mean is employed), if there is a transcript of
exactly the mean fragment length, a divide by zero
could occur causing a NAN TPM value.  This commit
fixes this bug.